refactor: UserService 추상화(데코레이터 패턴 적용)
