+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,48 +1,12 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import javax.sql.DataSource;
-import nextstep.jdbc.DataAccessException;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.TransactionStatus;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
 
-public class UserService {
+public interface UserService {
 
-    private final PlatformTransactionManager transactionManager;
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+    User findById(final long id);
 
-    public UserService(final DataSource dataSource, final UserDao userDao,
-                       final UserHistoryDao userHistoryDao) {
-        this.transactionManager = new DataSourceTransactionManager(dataSource);
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
+    void insert(final User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        final TransactionStatus transaction = transactionManager.getTransaction(new DefaultTransactionDefinition());
-        try {
-            final var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-            transactionManager.commit(transaction);
-        } catch (Exception exception) {
-            transactionManager.rollback(transaction);
-            throw new DataAccessException(exception.getMessage());
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
